refactor: Reorganize to fit existing architecture
